+++ python/examples/prim/simple_cubic_binary.py
@@ -43,4 +43,4 @@ for op in factor_group:
 
 # Format as JSON
 with open("../../doc/examples/prim/json/simple_cubic_binary.json", "w") as f:
-    f.write(prim.to_json())
+    f.write(xtal.pretty_json(prim.to_dict()))