+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to `libcasm-xtal` will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Make OccupantDoFIsEquivalent call operators non-const and remove mutable private members so that behvavior is clear for multithreaded use.
+
+
 ## [2.3.0] - 2025-12-15
 
 ### Fixed